LSR (alias of UBFM) in Data Processing Immediate. (#20)
### Description:
Add LSR instruction which is an alias of UBFM and is part of Bitfield
instructions of Data Processing - Immediate (DPI)

https://developer.arm.com/documentation/ddi0602/2023-12/Base-Instructions/LSR--immediate---Logical-Shift-Right--immediate---an-alias-of-UBFM-


### Testing:

`make all` including conformance testing.

#### Call out:
There is a specific rand function to add LSR immediate instructions.
However, the conformance test output shows mostly UBFX and UBFIZ and a
few LSR and LSL in the output. I think the specific function is not
being called.

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.