Start FIPS-aligned model of AES specification
This commit adds `AESSpec.lean` which aims to model the AES
specification in a way that's as close to the FIPS specification
as possible (the existing specification in `AESArm.lean` is more
aligned to the Arm ASL).

This is work in progress. For the moment, we only have definitions
of the AES state as a bitvector, byte vector, or byte array, as well as
conversions between them.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>