Use List.replicate instead of List.create (#36)
### Description:

This PR removes the definition of List.create:

`List.create (x : BitVec n) (len: Nat) : List (BitVec n)`

and instead uses `List.replicate` from the Lean library.

Due to better support in Lean for the length of the return value from
`List.replicate`, a lot of the `have` lemmas for reasoning about length
of `List.create` results are no longer needed. Therefore the code is
simplified.

Some of the test vectors in Tests/AES-GCM/AESGCMProgramTests.lean are
large and we are hitting default `maxRecDepth`, so it is set to higher
(512->2100) in that file.

### Testing:

`make all` and conformance testing was successful

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.