Use BitVec.cast instead of \t (#43)
### Description:

Use of `BitVec.cast` is preferred instead of `\t`. This PR rewrites uses
of `\t` to `BitVec.cast`.

In doing so, it causes the theorem `sha512h_rule_2` in
`Proofs/SHA512/Sha512_block_armv8_rules.lean` to fail. This is because
extra `BitVec.cast` is introduced into the goal that could not be
recognized by `bv_decide`. By simplifying the goal using rule
`BitVec.cast_eq`, `BitVec.cast` is removed from the goal and the theorem
goes through again.

### Testing:

"make all" and cosimulation on Graviton2 and Graviton3 succeed.

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.