Use bv_check instead of bv_decide for SHA512 proofs
We now store some LRAT certificates in the repo that are used by
bv_check. However, we have temporarily skipped two of the longest
running bv_decide proofs `sha512h_rule_1` and `sha512h_rule_2`, which
take ~5min and ~3min to be solved, respectively. This is because their
generated LRAT certificates are ~200MB and ~120MB respectively, which
exceeds Github file size limits.