Add tests for GCMV8 spec functions (#45)
### Description:

This PR adds tests for GCMV8 functions including `GCMInitV8`,
`GCMGmultV8` and `GCMGhashV8`. These serves as examples for showing the
relationship between the specification and the ARM assembly functions.

### Testing:

"make all" succeeds and cosimulation test succeeds on Graviton2 and
Graviton3 machines.

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.