feat: add CSE v1
While the actual CSE logic seems robust, the part of the code
that is't yet robust is one that deals with running CSE till fixpoint.
In particular, the problem is that when we use `generalize`, it (of course)
does not know about our `Expr`s, and thus future `generalize`s will not be correct.
What we should do is to then `kabstract` on our expressions to perform replacement.
This is pretty subtle, and will take a while to get right.