Add ELFLoader test for AES-GCM assembly routines (#101)
### Description:

This PR adds ELFLoader tests for AES-GCM assembly routines. 

When updating the binary file crypto_test, the file exceeds the maximum
file size on Github. In this PR, crypto_test is removed and it is now
built in the CI.

Building of the AWS-LC is done in two steps, first a docker container is
built for dependencies. Then this docker (for x86) is run on Ubuntu
(x86) and cross-compilation of AWS-LC is done within the docker
container to generate the binary of AWS-LC for `neoverse-n1` (AArch64).
Then the compiled results are copied to designated location in LNSym
directory.

PC relative addressing instructions and the add_sub_imm instructions
after it uses immediate values that could be changed when linking. To
solve this issue, we use the .S.o files for some of the ELF loader
tests. The .S.o files are before linking and use address zero.

A new Makefile target is created just for the ELF tests for AWS-LC. This
allows the expensive build to be separated from target `all`, allowing
local development to be more efficient.

### Testing:

What tests have been run? Did `make all` succeed for your changes? Was
conformance testing successful on an Aarch64 machine? Not yet

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.