feat: simp-based aggregation of effects (#133)
This PR is stacked upon #129 

### Description:

Change `AxEffect` and `SymContext` to track all axiomatic effect
hypothesis that are added during symbolic evaluation, and feed these to
a `simp` post-processing step to implement a basic form of effect
aggregation.
This replaces the manual `simp`ing we've been doing so far, and works
well enough for the examples we've been interested in so far, but we
might want to consider a more principled approach to effect aggregation
when we run into scalability issues.

Additionally, to make the aggregation work properly, we add a pass that
checks whether symbolic evaluation has reached the final state, and does
the appropriate substitution (replacing `sf` with the last intermediate
state we introduced).
* For now, this check is relatively simple, and only works when the
amount of steps to run was a concrete literal, that reached 0 after
simulation. We could integrate this with the `whileTac` approach and
accept a user-extensible tactic to prove that `runSteps = 0`.
* We add initial `h_program`, `h_pc` and `h_err` hypotheses to the
post-processing simp-set, so that goals like "the step after simulating
for n steps is error-free" will be closed entirely automatically by
calling `sym_n $n`
* We also expose a separate `sym_aggregate` tactic, which will do a
type-based search of the local context to determine (non-)effect
hypotheses. This works better when simulation is spread out over
multiple `sym_n` calls.
* Additionally, now that we have a good aggregation story, we no longer
expose `h_step` lemmas in the goal states, making those a tad less
cluttered

### Testing:

 `make all` ran locally

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.

---------

Co-authored-by: Siddharth <siddu.druid@gmail.com>
Co-authored-by: Shilpi Goel <shigoel@gmail.com>