feat: change `AxEffects.toSimpTheorems` to generate more specific discrtree keys (#148)
### Description:

The method we were using to construct simp-lemmas sets a `noIndexAtArgs`
option, meaning that a lemma like
`r (.GPR 0#5) s10` would get indexed *only* on `r`, and the arguments
were ignored. This caused a lot of attempts at unification.

Ideally, we'd want `mkSimpTheorems` to expose `noIndexAtArgs` as an
option, so we could disable it.
For now, we workaround the issue with a postprocessing step that
regenerates the discrtree keys with this behaviour turned off. Sadly,
the failed unifications turn out to not be the main expense: with this
change we can simulate SHA512 for 190 steps, which is a relatively small
improvement over before.

Then, I set the priority of the field-specfic lemmas higher than the
default priority used for the nonEffectProof, to prevent simp from
trying to discharge the expensive side-conditions when we already know a
cheaper lemma that is applicable. This did give an impressive speedup.

All together, we are now able to simulate 250 steps of SHA512 with the
default heartbeat limit (when tracing is turned off! Tracing does
consume some heartbeats, so with tracing the limit is lower)

### Testing:

`make all` succeeded locally

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.

---------

Co-authored-by: Shilpi Goel <shigoel@gmail.com>