Reason about one block of SHA512 (#174)
### Description:

We attempt to prove the functional correctness of SHA512 when exactly
one input block has to be hashed. The goal is to stress-test `sym_n`,
`simp_mem`, and `bv_decide`, and to determine whether we need other
kinds of automation.

### Testing:

What tests have been run? Did `make all` succeed for your changes? Was
conformance testing successful on an Aarch64 machine?

Yes.

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.