refactor: rename SymM to StepThmsM (#175)
### Description:

This monad is only used in step-theorem generation, so the new name is
more accurate, and frees up `SymM` to use as a monad to track
`SymContext` state.

### Testing:

What tests have been run? Did `make all` succeed for your changes? Was
conformance testing successful on an Aarch64 machine? Yes

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.

Co-authored-by: Shilpi Goel <shigoel@gmail.com>