Remove cadical build/install in workflows (#186)
### Description:

Cadical is no longer a user-facing dependency of LNSym, now that it is
available with Lean.

### Testing:

What tests have been run? Did `make all` succeed for your changes? Was
conformance testing successful on an Aarch64 machine?

Yes

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.