feat: Correctness of Memcpy program (#138)
We add a manual proof of `memcpy` to understand just how complex a proof
will be. The proof challenges `simp_mem`, due to, what we suspect, are
`omega` scaling bottlenecks. This file will be used as a litmus test for
improving automation.

### Testing:

`make all` succeeds, nothing was change with regards to conformance.

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.

---------

Co-authored-by: Shilpi Goel <shigoel@gmail.com>