refactor: cleanup popcount proof (#168)
### Description:

We've cleaned up the popcount proof by:
* extending `sym_aggregate` to also search for memory effect hypotheses
to simplify, and
* changing the semantics of `add_sub_immediate` to avoid
pattern-matching. It turns out that pattern-matching on the result of an
`if` (e.g., `let (a, b) := if x = 0 then (foo, bar) else (bar, foo)`)
get's reduced to a `Decidable.rec` by `simp only`, which then gets
stuck. By rephrasing it in terms of projections, we avoid this, and get
clean simplified semantics.
* In the same vein, we also add existing simplification rules for
`AddWithCarry` to the `bitvec_rules` simpset

### Testing:

`make all` succeeds.

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.

---------

Co-authored-by: Siddharth Bhat <siddu.druid@gmail.com>
Co-authored-by: Shilpi Goel <shigoel@gmail.com>