refactor: remove `SymContext.h_err?` field, move `addGoalsForMissingHypotheses` functionality into `fromLocalContext` (#181)
### Description:

Continues the cleanup started in #180 by getting rid of `h_err?` and
`addGoalsForMissingHypotheses`.

Note that we remove the ability to add a goal for `CheckSPAlignment`,
this option was false by default, and never used

### Testing:

What tests have been run? Did `make all` succeed for your changes? Was
conformance testing successful on an Aarch64 machine? yes

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.

Co-authored-by: Shilpi Goel <shigoel@gmail.com>