chore: delete bv_omega', reuse bv_omega (#194)
### Description:

We want to reuse as much of `bv_omega` as possible, and implement
`bv_omega` within `simp_mem` by calling the appropriate `simp` and
`omega` meta level functions, with the goal of extracting the
counter-example model from `omega` for lazy unfolding (ala CEGAR). To do
this, we first delete our layers on top of `bv_omega`. The next patch in
the series will replace the call of `evalTactic (tactic| bv_omega)` with
the appropriate meta-level functions. This will allow us to inspect the
state of `bv_omega`, instead of it being a black-box tactic invocation.

This is a chore PR.

### Testing:

What tests have been run? Did `make all` succeed for your changes? Was
conformance testing successful on an Aarch64 machine?

Conformance succeeds. 

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.