fix: `inferStatePrefix...` should default to `s1` as the first intermediate state. (#199)
### Description:

We claim in the error that we fall back to `s1` as the first
intermediate state if we can't infer anything.
However, the implementation actually used `s2`. 

This PR fixes that by setting `currentStateNumber := 0` in the fallback
case. The off-by-one behaviour is because this number is incremented
*before* it's used to generate a new state.

### Testing:

What tests have been run? Did `make all` succeed for your changes? Was
conformance testing successful on an Aarch64 machine? Yes

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.