feat: add combinators for more fine-grained benchmarking (#198)
### Description:

This adds combinators to benchmark specific parts of a tactic, either by
directly outputting the data, or by storing datapoints in a state for
aggregation. Either behaviour is controlled by a `benchmark` option,
which defaults to false, but is set to `true` in the benchmark command.

### Testing:

What tests have been run? Did `make all` succeed for your changes? Was
conformance testing successful on an Aarch64 machine? yes

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.

---------

Co-authored-by: Shilpi Goel <shigoel@gmail.com>