remove `profilerDir` option
The `trace.profiler.output` option has to be set when the frontend finishes the entire file. Therefore, it doesn't make sense to control it from within the benchmark elaborator: it's likely to do unexpected things if we have multiple benchmarks in the same file.

Instead, we should be setting the option from the makefile that runs the benchmark to begin with