Refactor to use `extractLsb'` instead of `extractLsb` (#191)
## Description:

This PR refactors to use `extractLsb'` instead of `extractLsb` in the
whole repository. Doing so, we are able to simplify the proofs and
remove unnecessary hypothesis inputs to several functions.

### Testing:

What tests have been run? Did `make all` succeed for your changes? Was
conformance testing successful on an Aarch64 machine? Yes.

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.