refactor: use searchLCtx in sym_aggregate (#201)
### Description:

Stacked on #200.

This uses the `searchLCtx` machinery we built in #189 in
`sym_aggregate`'s implementation.
For now, this searches for exactly the same kind of expressions as
before, but this makes it much easier to expand this set in the next PR.

### Testing:

What tests have been run? Did `make all` succeed for your changes? Was
conformance testing successful on an Aarch64 machine? yes

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.

---------

Co-authored-by: Shilpi Goel <shigoel@gmail.com>