feat: add bv_omega_bench to run omega and produce goal states + timing inf. (#223)
### Description:

This PR replaces `bv_omega` with `bv_omega_bench`, which is used to
write benchmarking results into a user-specified filepath. This enables
us to extract out benchmarks to be upstreamed, as begun in
https://github.com/leanprover/lean4/pull/5622.

We make the file path, whether the benchmark run is enabled or not, and
the minimum time necessary to be added to the benchmark all
user-configurable parameters.


### Testing:

What tests have been run? Did `make all` succeed for your changes? Was
conformance testing successful on an Aarch64 machine?

### License:

By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.

---------

Co-authored-by: Shilpi Goel <shigoel@gmail.com>